refactor: split into more files according to nextauth v5
